Revert changes to fan speed.
Setting the fan LUT to start at target_temp has the desired effect of
starting the fan after the temperature has been reached but the
undesired effect of keeping it on until the temperature drops 1 degree
below this setting. Since we're trying to maintain a minimum chamber
temperature it's more important that the fan turns OFF at that setting
so use target_temp + 1 as we had before.

This reverts commit 5f1ffb3676e2b58ff0481196aedf343500d49cfe.